removed (apparently) useless +version method from classes, due to problems with the ObjC2 runtime on OpenBSD (thanks to Sebastian Reitenbach)

+++ sope-core/NGExtensions/NGBundleManager.m
@@ -615,7 +615,7 @@ static NSString *NGEnvVarPathSeparator = @":";
 
 // dependencies
 
-+ (NSInteger)version {
++ (int)version {
   return 2;
 }
